修复web传入倍率是小数，导致解析成整数失败的问题

+++ gms-server/src/main/java/org/gms/model/dto/GiveResourceReqDTO.java
@@ -13,7 +13,7 @@ public class GiveResourceReqDTO {
     private Byte type;
     private Integer id;
     private Integer quantity;
-    private Float rate;
+    private Integer rate;
     private Short str;
     private Short dex;
     @JsonProperty("int")
